+++ sema4ai/vscode-client/src/robo/agentPackage.ts
@@ -9,6 +9,7 @@ import {
     getWorkspacePackages,
     PackageTargetAndNameResult,
     revealInExtension,
+    showErrorMessageWithShowOutputButton,
 } from "../common";
 import { ActionResult, LocalPackageMetadataInfo, PackageType, PackageYamlName } from "../protocols";
 import * as roboCommands from "../robocorpCommands";
@@ -98,8 +99,7 @@ export const importAgentPackage = async (): Promise<void> => {
     } catch (err) {
         const errorMsg = "Error importing Agent Package. Target: " + targetDir + " File: " + uriResult;
         logError(errorMsg, err, "ERR_IMPORT_ACTION_PACKAGE");
-        OUTPUT_CHANNEL.appendLine(errorMsg);
-        window.showErrorMessage(errorMsg + " (see `OUTPUT > Sema4.ai` for more details).");
+        showErrorMessageWithShowOutputButton(errorMsg);
     }
 };
 
@@ -133,8 +133,7 @@ export const createAgentPackage = async (): Promise<void> => {
     } catch (err) {
         const errorMsg = "Error creating Agent Package at: " + targetDir;
         logError(errorMsg, err, "ERR_CREATE_ACTION_PACKAGE");
-        OUTPUT_CHANNEL.appendLine(errorMsg);
-        window.showErrorMessage(errorMsg + " (see `OUTPUT > Sema4.ai` for more details).");
+        showErrorMessageWithShowOutputButton(errorMsg);
     }
 };
 
@@ -153,10 +152,7 @@ export const packAgentPackage = async (targetDir: string): Promise<{ zipPath: st
     } catch (error) {
         const errorMsg = `Failed to package the agent at: ${targetDir}`;
         logError(errorMsg, error, "ERR_PACK_ACTION_PACKAGE");
-
-        const detailedErrorMsg = `${errorMsg}. Please check the 'OUTPUT > Sema4.ai' for more details.`;
-        OUTPUT_CHANNEL.appendLine(detailedErrorMsg);
-        window.showErrorMessage(detailedErrorMsg);
+        showErrorMessageWithShowOutputButton(errorMsg);
     }
 
     return null;
@@ -252,10 +248,7 @@ export const updateAgentVersion = async (agentPath: string): Promise<void> => {
     } catch (error) {
         const errorMsg = `Failed to update the agent at: ${agentPath}`;
         logError(errorMsg, error, "ERR_UPDATE_AGENT_VERSION");
-
-        const detailedErrorMsg = `${errorMsg}. Please check the 'OUTPUT > Sema4.ai' for more details.`;
-        OUTPUT_CHANNEL.appendLine(detailedErrorMsg);
-        window.showErrorMessage(detailedErrorMsg);
+        showErrorMessageWithShowOutputButton(errorMsg);
     }
 };
 
@@ -286,9 +279,6 @@ export const refreshAgentSpec = async (agentPath: string): Promise<void> => {
     } catch (error) {
         const errorMsg = `Failed to refresh the Agent Configuration`;
         logError(errorMsg, error, "ERR_REFRESH_AGENT_SPEC");
-
-        const detailedErrorMsg = `${errorMsg}. Please check the 'OUTPUT > Sema4.ai' for more details.`;
-        OUTPUT_CHANNEL.appendLine(detailedErrorMsg);
-        window.showErrorMessage(detailedErrorMsg);
+        showErrorMessageWithShowOutputButton(errorMsg);
     }
 };